+++ fixtures/output-app/config.rb
@@ -1,0 +1,4 @@
+activate :google_analytics do |ga|
+  ga.tracking_id = 'UA-123456-78'
+  ga.output = :js
+end