+++ lib/middleman-google-analytics/extension.rb
@@ -2,6 +2,7 @@ require 'middleman-google-analytics/helpers'
 
 module Middleman
   class GoogleAnalyticsExtension < Extension
+
     option :tracking_id, nil, 'Property ID'
     option :anonymize_ip, false, 'Removing the last octet of the IP address'
     option :domain_name, nil, 'Tracking across a domain and its subdomains'
@@ -10,23 +11,24 @@ module Middleman
     option :debug, false, 'Tracking Code Debugger'
     option :development, true, 'Tracking in development environment'
     option :minify, false, 'Compress the JavaScript code'
-
-    def initialize(app, options_hash={}, &block)
-      super
-
-      app.set :google_analytics_settings, options
-    end
+    option :output, :html, 'Output style - :html includes <script> tag'
 
     def after_configuration
       unless options.tracking_id
         $stderr.puts 'Google Analytics: Please specify a property ID'
-        raise 'No property ID given' if display?
+        raise ArgumentError, 'No property ID given' if display?
       end
 
       if options.allow_linker and not options.domain_name
         $stderr.puts 'Google Analytics: Please specify a domain_name when ' \
                      'using allow_linker'
-        raise 'No domain_name given' if display?
+        raise ArgumentError, 'No domain_name given' if display?
+      end
+
+      unless [:html, :js].include?(options.output.try(:to_sym))
+        $stderr.puts 'Google Analytics: Please specify a valid output ' \
+                     'type (html|js).'
+        raise ArgumentError, 'Only "html" or "js" allowed' if display?
       end
     end