+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-## HEAD (???)
+## 2.0.0 (2015-12-16)
+
+* Middleman v4 support added.
+* Global google_analytics_settings removed.
+* Output option added.
+* Erubis instead of ERB.
+* Require Ruby 2.0 or greater.
+* Require Middleman 3.4 or greater.
 
 ## 1.1.0 (2015-01-06)